requirements.txt
asgiref==3.8.1
Django==5.1.2
django-countries==7.6.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.48
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
